+++ next.config.js
@@ -1,5 +1,7 @@
 /** @type {import('next').NextConfig} */
-require('reflect-metadata')
-module.exports = {
+require('reflect-metadata');
+const withLess = require("next-with-less");
+module.exports = withLess({
   reactStrictMode: true,
-}
+  lessLoaderOptions: {},
+})