+++ src/pages/_app.tsx
@@ -1,7 +1,7 @@
 import { Layout } from '@/components/layout';
 import { SampleApiClientProvider } from '@/hooks/sample-api/sample-api-client-hook';
 import '@/styles/globals.css';
-import 'antd/dist/antd.css';
+import '@/styles/antd.less';
 import type { AppProps } from 'next/app';
 
 function MyApp({ Component, pageProps }: AppProps) {